requirements.txt
torch
torchvision
numpy
scipy
matplotlib
pillow
scikit-learn
streamlit
opencv-python
pandas